updates
I realized I still need to implement a feature in the new file structure to do the segmentation step on a batch of point clouds. I don't have time to do that today, so I will have to do that on Monday. That means that the "Everything" tool and batch processing for the "Segment" tool currently are not working. You can still do batch processing for the "Clean" tool and single point cloud processing for the "Segment" Tool. The "Sorting" tool is also fully working.

Co-Authored-By: alexanderswann <32579236+alexanderswann@users.noreply.github.com>